rename standardBehaviors to standardEvents and move to scripts file
